+++ apps/prairielearn/src/lib/enrollment/barrier.test.sql
